+++ src/jaxsim/rbda/__init__.py
@@ -1,11 +1,8 @@
 StandardGravity = 9.81
 
-from jaxsim.physics.algos import (
-    aba,
-    crba,
-    forward_kinematics,
-    jacobian,
-    rnea,
-    soft_contacts,
-    utils,
-)
+from .aba import aba
+from .crba import crba
+from .forward_kinematics import forward_kinematics, forward_kinematics_model
+from .jacobian import jacobian
+from .rnea import rnea
+from .soft_contacts import SoftContacts, SoftContactsParams, collidable_points_pos_vel
